Update es.json
Updated and double checked spell in Spanish (SP)- Works for Spain and Latin America. There was missing translations and there are some terms that should not be translated as the Swap titles ( Atomic Swap)  or "billetera" , since they are wide used anglicism and translate them makes them go out of context.